fix: dependencies to work on snelllius, w&b logging, update README to install working venv in cluster
