+++ data/segthor_train.sha256
