+++ requirements.txt
@@ -25,7 +25,7 @@ huggingface-hub==0.25.0
 idna==3.8
 imageio==2.35.1
 ipykernel==6.29.5
-ipython==8.27.0
+ipython==8.18.1
 itk==5.4.0
 itk-core==5.4.0
 itk-elastix==0.21.0
@@ -51,16 +51,16 @@ monai==1.3.2
 mpmath==1.3.0
 multidict==6.1.0
 nest-asyncio==1.6.0
-networkx==3.3
+networkx==3.2.1
 nib==0.5.7
 nibabel==5.2.1
-numpy==1.26.4
+numpy==1.24.4
 packaging==24.1
 parso==0.8.4
 peft==0.12.0
 pexpect==4.9.0
 pillow==10.4.0
-platformdirs==4.3.2
+platformdirs==3.9.1
 prompt-toolkit==3.0.47
 protobuf==5.28.1
 psutil==6.0.0
@@ -76,7 +76,7 @@ regex==2024.9.11
 requests==2.32.3
 safetensors==0.4.5
 scikit-image==0.24.0
-scipy==1.14.1
+scipy==1.13.1
 sentry-sdk==2.14.0
 setproctitle==1.3.3
 setuptools==74.1.2
@@ -88,15 +88,15 @@ stack-data==0.6.3
 sympy==1.13.2
 tifffile==2024.8.30
 tokenizers==0.19.1
-torch==2.2.2
+torch==2.1.0
 torchmetrics==1.4.2
-torchvision==0.17.2
+torchvision==0.16.0
 tornado==6.4.1
 tqdm==4.66.5
 traitlets==5.14.3
 transformers==4.44.2
 typing-extensions==4.12.2
-urllib3==2.2.3
+urllib3==1.26.16
 wandb==0.18.0
 wcwidth==0.2.13
 yarl==1.11.1
